Added house scene, finished furniture scripts, started level 3 scripts
So I finished the furniture scripts and tied it to a UI that will update dynamically as the inventory is updated in each scene. I added level 3 and started adding scripts for some of the behaviors we will need. I also organized some of the scripts by level so they will be easier to find and use.